Add a helper function in ListModule, reverse().
Chapter 3 Exercise 5: ForeachExample prints the arguments in reverse
order. Determine the cause and implement a fix. Hint: consider adding a
usefule method to ListHelper that is commonly found in List classes.

The cause that arguments are printed in reverse order is that the
temporary list where arguments are added becomes tail of next temporary
list.

A helper function, reverse(), is implemented in ListModule instead of
ListHelper.